LayoutTests/fast/text/line-breaks-after-ideographic-comma-or-full-stop-expected.txt

These are good:
う、
texttexttexttext

う。
texttexttexttext

The following two should look like “good”:
う、texttexttexttext

う。texttexttexttext

These are bad:
う、texttexttexttext

う。texttexttexttext

PASS
